+++ src/main/java/subway/ErrorView.java
@@ -1,0 +1,4 @@
+package subway;
+
+public class ErrorView {
+}
